# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktaba Dot Com - Arabic RTL E-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginCustomer
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApplyInput"
      responses:
        "200":
          description: Cart with coupon applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCoupon
      tags: [cart]
      responses:
        "200":
          description: Cart without coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── ORDERS (PUBLIC) ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    CustomerRegisterInput:
      type: object
      required: [name, mobile, password]
      properties:
        name: { type: string }
        email: { type: ["string", "null"] }
        mobile: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [customer]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
    AdminAuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"

    # ── Customer ─────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, mobile, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        mobile: { type: string }
        isBlocked: { type: boolean }
        createdAt: { type: string }
...
    AdminUser:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        permissions: { type: array, items: { type: string } }

    # ── Address ──────────────────────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [productId, quantity, unitPrice, subtotal]
      properties:
        productId: { type: integer }
        nameAr: { type: string }
        coverImage: { type: ["string", "null"] }
        slug: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        oldPrice: { type: ["number", "null"] }
        subtotal: { type: number }
        inStock: { type: boolean }
        stockQuantity: { type: integer }
    Cart:
      type: object
      required: [items, subtotal, shippingCost, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        discount: { type: number }
        shippingCost: { type: number }
        total: { type: number }
        couponCode: { type: ["string", "null"] }
        couponDiscount: { type: number }
        governorateId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
    CouponApplyInput:
      type: object
      required: [code]
      properties:
        code: { type: string }

    # ── Orders ───────────────────────────────────────────────────────────────
...